move mergeSchemas from utils to Service static method
