+++ src/utils.js
@@ -77,58 +77,11 @@ let utils = {
 		return (p != null && typeof p.then === "function");
 	},
 
-
 	/**
-	 * Merge two Service schema
+	 * Clear `require` cache. Used for service hot reloading
 	 *
-	 * @param {Object} schema
-	 * @param {Object} mods
-	 * @returns
+	 * @param {String} filename
 	 */
-	mergeSchemas(schema, modSchema) {
-		function updateProp(propName, target, source) {
-			if (source[propName] !== undefined)
-				target[propName] = source[propName];
-		}
-
-		const res = _.cloneDeep(schema);
-		const mods = _.cloneDeep(modSchema);
-
-		Object.keys(mods).forEach(key => {
-			if (["settings", "metadata"].indexOf(key) !== -1) {
-				res[key] = _.defaultsDeep(mods[key], res[key]);
-			} else if (["actions"].indexOf(key) !== -1) {
-				if (res[key] == null)
-					res[key] = {};
-
-				Object.keys(mods[key]).forEach(k => {
-					res[key][k] = _.defaultsDeep(
-						_.isFunction(mods[key][k]) ? { handler: mods[key][k] } : mods[key][k],
-						_.isFunction(res[key][k]) ? { handler: res[key][k] } : res[key][k]
-					);
-				});
-			} else if (["methods"].indexOf(key) !== -1) {
-				res[key] = _.assign(res[key], mods[key]);
-			} else if (["events"].indexOf(key) !== -1) {
-				if (res[key] == null)
-					res[key] = {};
-
-				Object.keys(mods[key]).forEach(k => {
-					res[key][k] = _.compact(_.flatten([res[key][k], mods[key][k]]));
-				});
-			} else if (["created", "started", "stopped"].indexOf(key) !== -1) {
-				// Concat lifecycle event handlers
-				res[key] = _.compact(_.flatten([res[key], mods[key]]));
-			} else if (["mixins", "dependencies"].indexOf(key) !== -1) {
-				// Concat mixins
-				res[key] = _.compact(_.flatten([mods[key], res[key]]));
-			} else
-				updateProp(key, res, mods);
-		});
-
-		return res;
-	},
-
 	clearRequireCache(filename) {
 		/* istanbul ignore next */
 		Object.keys(require.cache).forEach(function(key) {
